+++ app/library/System/Database/MyQuery.php
@@ -34,7 +34,7 @@ class MyQuery
     return new Table($table_name, $this->PDO);
   }
 
-  public static function conn(string $table_name, MyPDO $PDO = null)
+  public static function from(string $table_name, MyPDO $PDO = null)
   {
     $conn = new MyQuery($PDO);
     return $conn->table($table_name);