+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- MyQuery -> update support add mulity values using single function
+
+### Changed
+- MyQuery change function name from conn() to from()
+
+### Fixed
+- MyQuery -> insert, remove var_dump() in production
+- Mymodel -> cant add value 0 (zero)
+- Mymodel -> remove where statment if no bind query
 
 ## [0.1.4] - 2021-04-23
 ### Added